fix stock sender

+++ src/Integration/Ozon/Clients/Stocks/Messages/OzonStockRequest.cs
@@ -11,9 +11,9 @@ public class OzonStock
 {
     [JsonProperty("offer_id")] public string OfferId { get; set; }
 
-    [JsonProperty("product_id")] public int? ProductId { get; set; }
+    [JsonProperty("product_id")] public long? ProductId { get; set; }
 
     [JsonProperty("stock")] public int Stock { get; set; }
-    
+
     [JsonProperty("warehouse_id")] public long WarehouseId { get; set; }
 }
