+++ src/Integration/Ozon/Services/OzonSendStockService.cs
@@ -26,22 +26,22 @@ public class OzonSendStockService : MarketplaceSendStockService
         var client = ServiceProvider.GetRequiredService<IOzonStockClient>();
         var logger = ServiceProvider.GetRequiredService<ILogger<OzonSendStockService>>();
         var context = ServiceProvider.GetRequiredService<MContext>();
-        
+
         var ozon = Mapper.Map<OzonDto>(marketplace);
 
         var warehouses = await context.OzonWarehouses
             .AsNoTracking()
             .Where(x => x.OzonId == ozon.Id)
-            .Select(x => new {x.OzonWarehouseId, x.Id})
+            .Select(x => new { x.OzonWarehouseId, x.Id })
             .ToListAsync();
 
         var warehouseIds = warehouses.Select(x => x.Id);
-        
+
         var blacklists = (await context.OzonWarehouseBlacklists
-            .AsNoTracking()
-            .Where(x => warehouseIds.Contains(x.OzonWarehouseId))
-            .Select(x => new {x.OzonWarehouseId, x.ProductId})
-            .ToListAsync())
+                .AsNoTracking()
+                .Where(x => warehouseIds.Contains(x.OzonWarehouseId))
+                .Select(x => new { x.OzonWarehouseId, x.ProductId })
+                .ToListAsync())
             .GroupBy(x => x.OzonWarehouseId)
             .ToDictionary(x => x.Key, x => x
                 .Select(bl => bl.ProductId)
@@ -53,21 +53,21 @@ public class OzonSendStockService : MarketplaceSendStockService
         {
             Stocks = new List<OzonStock>()
         };
-        
+
         requests.Add(currentRequest);
-        
+
         var stocksWithExternalId = stocks
             .Where(s => !s.ProductExternalId.IsNullOrEmpty());
 
         foreach (var stock in stocksWithExternalId)
         {
-            if (!int.TryParse(stock.ProductExternalId, out var ozonProductId))
+            if (!long.TryParse(stock.ProductExternalId, out var ozonProductId))
             {
                 throw new Exception($"Wrong ozon external id: {stock.ProductExternalId} for product {stock.ProductId}");
             }
 
             var stockValue = Convert.ToInt32(stock.Value);
-            
+
             var ozonStocks = warehouses.Select(warehouse =>
             {
                 var warehouseStock = stockValue;
@@ -77,6 +77,7 @@ public class OzonSendStockService : MarketplaceSendStockService
                     logger.LogInformation("Product ${ProductId} is blacklisted for warehouse ${WarehouseId}",
                         stock.ProductId, warehouse.OzonWarehouseId);
                 }
+
                 return new OzonStock
                 {
                     Stock = warehouseStock,
